backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
openai==1.54.0
supabase==2.10.0
langsmith>=0.1.0
pydantic==2.10.3
pydantic-settings==2.7.0
sse-starlette==1.8.2
tiktoken==0.7.0
python-multipart==0.0.9
docling
sentence-transformers
tavily-python
